requirements.txt
xmpppy==0.6.1